// frontend/src/types/filterDateType.ts
export type BtnType = {
    FILTER_DAY: HTMLElement | null,
    FILTER_WEEK: HTMLElement | null,
    FILTER_MONTH: HTMLElement | null,
    FILTER_YEAR: HTMLElement | null,
    FILTER_ALL: HTMLElement | null,
    FILTER_INTERVAL: HTMLElement | null
}

export type PeriodHandlersType = {
    FILTER_ALL: string;
    FILTER_DAY: string;
    FILTER_WEEK: string;
    FILTER_MONTH: string;
    FILTER_YEAR: string;
    FILTER_INTERVAL: string;
}